Refactor previous graphl files
