+++ web/src/pages/AlertDetails/graphql/alertDetails.graphql
@@ -9,15 +9,3 @@ query AlertDetails($input: GetAlertInput!) {
         events
     }
 }
-
-query RuleTeaser($input: GetRuleInput!) {
-    rule(input: $input) {
-        description
-        displayName
-        id
-        logTypes
-        runbook
-        severity
-        tags
-    }
-}
