+++ web/src/pages/AlertDetails/graphql/ruleTeaser.graphql
@@ -1,0 +1,11 @@
+query RuleTeaser($input: GetRuleInput!) {
+    rule(input: $input) {
+        description
+        displayName
+        id
+        logTypes
+        runbook
+        severity
+        tags
+    }
+}
