+++ web/src/pages/AlertDetails/index.tsx
@@ -19,3 +19,4 @@
 export { default } from './AlertDetails';
 export * from './AlertDetails';
 export * from './graphql/alertDetails.generated';
+export * from './graphql/ruleTeaser.generated';
