+++ web/src/pages/ComplianceOverview/graphql/getOrganizationStats.generated.ts
