+++ web/src/pages/ComplianceOverview/graphql/getOrganizationStats.graphql
