+++ web/src/pages/CreateComplianceSource/CreateComplianceSource.tsx
@@ -22,12 +22,12 @@ import { Card, Flex, Alert, Box } from 'pouncejs';
 import { INTEGRATION_TYPES, AWS_ACCOUNT_ID_REGEX } from 'Source/constants';
 import urls from 'Source/urls';
 import { extractErrorMessage } from 'Helpers/utils';
-import { ListInfraSourcesDocument } from 'Pages/ListComplianceSources/ComplianceSourceTable/graphql/complianceSourceTable.generated';
+import { ListInfraSourcesDocument } from 'Pages/ListComplianceSources/ComplianceSourceTable/graphql/listInfraSources.generated';
 import useRouter from 'Hooks/useRouter';
 import { Formik } from 'formik';
 import * as Yup from 'yup';
 import { Wizard, WizardPanelWrapper } from 'Components/Wizard';
-import { useAddInfraSource } from 'Pages/CreateComplianceSource/graphql/createComplianceSource.generated';
+import { useAddInfraSource } from 'Pages/CreateComplianceSource/graphql/addInfraSource.generated';
 import RemediationPanel from './RemediationPanel';
 import RealTimeEventPanel from './RealTimeEventPanel';
 import ResourceScanningPanel from './ResourceScanningPanel';
