+++ web/src/pages/CreateComplianceSource/graphql/addInfraSource.generated.ts
