+++ web/src/pages/CreateComplianceSource/graphql/addInfraSource.graphql
