+++ web/src/pages/CreateLogSource/graphql/addLogSource.generated.ts
