+++ web/src/pages/CreateLogSource/graphql/addLogSource.graphql
