+++ web/src/pages/Destinations/index.tsx
@@ -18,4 +18,4 @@
 
 export { default } from './Destinations';
 export * from './Destinations';
-export * from './graphql/destinations.generated';
+export * from './graphql/listDestinationsAndDefaults.generated';
