+++ web/src/pages/EditPolicy/graphql/updatePolicy.graphql
@@ -1,27 +1,3 @@
-query PolicyDetails($input: GetPolicyInput!) {
-    policy(input: $input) {
-        autoRemediationId
-        autoRemediationParameters
-        description
-        displayName
-        enabled
-        suppressions
-        id
-        reference
-        resourceTypes
-        runbook
-        severity
-        tags
-        body
-        tests {
-            expectedResult
-            name
-            resource
-            resourceType
-        }
-    }
-}
-
 mutation UpdatePolicy($input: CreateOrModifyPolicyInput!) {
     updatePolicy(input: $input) {
         autoRemediationId
