+++ web/src/pages/EditRule/graphql/ruleDetails.graphql
@@ -1,0 +1,20 @@
+query RuleDetails($input: GetRuleInput!) {
+    rule(input: $input) {
+        description
+        displayName
+        enabled
+        id
+        reference
+        logTypes
+        runbook
+        severity
+        tags
+        body
+        tests {
+            expectedResult
+            name
+            resource
+            resourceType
+        }
+    }
+}
