+++ web/src/pages/EditRule/graphql/updateRule.graphql
@@ -1,24 +1,3 @@
-query RuleDetails($input: GetRuleInput!) {
-    rule(input: $input) {
-        description
-        displayName
-        enabled
-        id
-        reference
-        logTypes
-        runbook
-        severity
-        tags
-        body
-        tests {
-            expectedResult
-            name
-            resource
-            resourceType
-        }
-    }
-}
-
 mutation UpdateRule($input: CreateOrModifyRuleInput!) {
     updateRule(input: $input) {
         description
