+++ web/src/pages/GeneralSettings/graphql/getGeneralSettings.graphql
@@ -1,0 +1,7 @@
+query GetGeneralSettings {
+    generalSettings {
+        displayName
+        email
+        errorReportingConsent
+    }
+}
