+++ web/src/pages/GeneralSettings/graphql/updateGeneralSettings.graphql
@@ -1,11 +1,3 @@
-query GetGeneralSettings {
-    generalSettings {
-        displayName
-        email
-        errorReportingConsent
-    }
-}
-
 mutation UpdateGeneralSettings($input: UpdateGeneralSettingsInput!) {
     updateGeneralSettings(input: $input) {
         displayName
