+++ web/src/pages/ListComplianceSources/ComplianceSourceTable/graphql/listInfraSources.generated.ts
