+++ web/src/pages/ListComplianceSources/ComplianceSourceTable/graphql/listInfraSources.graphql
