+++ web/src/pages/ListComplianceSources/index.tsx
@@ -18,4 +18,4 @@
 
 export { default } from './ListComplianceSources';
 export * from './ComplianceSourceTable';
-export * from './ComplianceSourceTable/graphql/complianceSourceTable.generated';
+export * from './ComplianceSourceTable/graphql/listInfraSources.generated';
