+++ web/src/pages/ListLogSources/LogSourceTable/graphql/listLogSources.graphql
