+++ web/src/pages/ListResources/ListResourcesActions/graphql/listAccountIds.graphql
