+++ web/src/pages/ListResources/index.tsx
@@ -19,4 +19,4 @@
 export { default } from './ListResources';
 
 export * from './graphql/listResources.generated';
-export * from './ListResourcesActions/graphql/listResourcesActions.generated';
+export * from './ListResourcesActions/graphql/listAccountIds.generated';
